+++ generators/mock.rb
@@ -6,7 +6,10 @@ class Mock
   EXCLUDE_FROM_ALL = false
 
   def self.generate(file)
-    puts "Not overwriting file Mock#{file}.hpp.lit" and return if File.exists? "#{PATH}/Mock#{file}.hpp.lit"
+    if File.exists? "#{PATH}/Mock#{file}.hpp.lit"
+      puts "Not overwriting file #{PATH}/Mock#{file}.hpp.lit"
+      return
+    end 
     
     File.open("#{PATH}/Mock#{file}.hpp.lit", "w") do |f|
       f.write <<-EOS
